# ProductCatalog/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Desafio FastTrack - Catálogo de Produtos com Java e Spring Boot",
    "version": "1.0.0",
    "title": "Products Catalog API"
  },
  "host": "localhost:9999",
  "basePath": "/",
  "tags": [
    {
      "name": "product-controller",
      "description": "Product Controller"
    }
  ],
  "paths": {
    "/products": {
      "get": {
        "tags": [
          "product-controller"
        ],
        "summary": "findAll",
        "operationId": "findAllUsingGET",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "product-controller"
        ],
        "summary": "post",
        "operationId": "postUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "form",
            "description": "form",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductForm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      }
    },
    "/products/search": {
      "get": {
        "tags": [
          "product-controller"
        ],
        "summary": "findFiltered",
        "operationId": "findFilteredUsingGET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "max_price",
            "in": "query",
            "description": "max_price",
            "required": false,
            "type": "number",
            "format": "double"
          },
          {
            "name": "min_price",
            "in": "query",
            "description": "min_price",
            "required": false,
            "type": "number",
            "format": "double"
          },
          {
            "name": "q",
            "in": "query",
            "description": "q",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "product-controller"
        ],
        "summary": "findById",
        "operationId": "findByIdUsingGET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "product-controller"
        ],
        "summary": "update",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "form",
            "description": "form",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductForm"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "product-controller"
        ],
        "summary": "delete",
        "operationId": "deleteUsingDELETE",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    }
  },
  "definitions": {
    "Product": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "type": "number",
          "format": "double"
        }
      },
      "title": "Product"
    },
    "ProductForm": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "type": "number",
          "format": "double"
        }
      },
      "title": "ProductForm"
    }
  }
}